requirements.txt
absl-py==1.2.0
asttokens==2.0.5
attrs==22.1.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
coloredlogs==15.0.1
comet-ml==3.31.12
configobj==5.0.6
cycler==0.11.0
cysignals==1.11.2
Cython==0.29.32
decorator==5.1.1
docker-pycreds==0.4.0
dulwich==0.20.46
efficientnet-pytorch==0.7.1
everett==3.0.0
executing==0.8.3
fasttext==0.6.0
filelock==3.7.1
focal-loss-torch==0.0.9
fonttools==4.34.4
future==0.18.2
gensim==4.2.0
gif==3.0.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.47.0
huggingface-hub==0.8.1
humanfriendly==10.0
idna==3.3
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.15.0
kaleido==0.2.1
kiwisolver==1.4.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
memory-profiler==0.60.0
munch==2.5.0
neo4j==4.4.5
nltk==3.7
numexpr==2.8.3
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
plotly==5.9.0
pretrainedmodels==0.7.4
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.1
pyfasttext==0.4.6
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
segmentation-models-pytorch==0.3.0
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.96
sentry-sdk==1.7.2
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
sklearn==0.0
smart-open==6.2.0
smmap==5.0.0
smote==0.1
stack-data==0.3.0
tables==3.7.0
tenacity==8.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
timm==0.4.12
tokenizers==0.12.1
torch==1.12.0
torch-cluster==1.6.0
torch-geometric==2.0.4
torch-scatter==2.0.9
torch-sparse==0.6.14
torch-spline-conv==1.2.1
torchaudio==0.12.0
torchmetrics==0.9.3
torchvision==0.13.0
tqdm==4.64.0
traitlets==5.3.0
transformers==4.21.0
typing-extensions==4.3.0
urllib3==1.26.10
wandb==0.12.21
wcwidth==0.2.5
websocket-client==1.3.3
Werkzeug==2.2.1
wrapt==1.14.1
wurlitzer==3.0.2
xgboost==1.6.2
zipp==3.8.1
